Upon trigger, run a treecompose.
We fork(), start a log file, run git-pull, then treecompose.
We log (to systemd) the start and completion.

This is kind of big, and in many ways ugly: it hardcodes a lot
of knowledge (git, treecompose commands, path to log file). It's
not terribly flexible. But that hardcoding has to happen somewhere,
and the alternative is to duplicate all that in multiple places
on the build server. This at least gives us consistency, helpful
log files, and a mechanism (journald) such that other processes
can trigger upon successful completion of a compose.